HDA_Project-main/requirements.txt
datasets
transformers
torch
scikit-learn
matplotlib
pandas
numpy
seaborn
umap-learn
tqdm
time
gensim
pathlib
